Module Impl.As_prover

The functions in this module may only be run as the prover; trying to run them outside of functions that refer to As_prover.t will result in a runtime error.

type 'a t = 'a

This type marks function arguments that can include function calls from this module. Using these functions outside of these will result in a runtime error.

type 'a as_prover = 'a t

val in_prover_block : unit -> bool

val read_var : Field.t -> Field.Constant.t

val read : ('var, 'value) Typ.t -> 'var -> 'value

include Snarky_intf.Field.Extended with type t := field

include Snarky_intf.Field.S with type t := field

include Ppx_sexp_conv_lib.Sexpable.S with type t := field

val t_of_sexp : Sexplib0__.Sexp.t -> field

val sexp_of_t : field -> Sexplib0__.Sexp.t

include Bin_prot.Binable.S with type t := field

val bin_size_t : field Bin_prot.Size.sizer

val bin_write_t : field Bin_prot.Write.writer

val bin_read_t : field Bin_prot.Read.reader

val __bin_read_t__ : (int -> field) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : field Bin_prot.Type_class.writer

val bin_reader_t : field Bin_prot.Type_class.reader

val bin_t : field Bin_prot.Type_class.t

val of_int : int -> field

val of_string : string -> field

val one : field

val zero : field

val add : field -> field -> field

val sub : field -> field -> field

val mul : field -> field -> field

val inv : field -> field

val square : field -> field

val sqrt : field -> field

val is_square : field -> bool

val equal : field -> field -> bool

val size_in_bits : int

val print : field -> unit

val random : unit -> field

module Vector : Snarky_intf.Vector.S with type elt = field

val to_string : field -> string

val negate : field -> field

val (+) : field -> field -> field

val (*) : field -> field -> field

val (-) : field -> field -> field

val (/) : field -> field -> field

val unpack : field -> bool list

Convert a field element into its constituent bits.

val project : bool list -> field