Module Impl.Typ

Mappings from OCaml types to R1CS variables and constraints.

module Store : sig ... end

module Alloc : sig ... end

module Read : sig ... end

type ('var, 'value) t = ('var, 'value, field, (unit, unit, field) Snarky_backendless.Checked.t) Snarky_backendless.Types.Typ.t

Accessors for Types.Typ.t fields:

val store : ('var, 'value) t -> 'value -> 'var Store.t

val read : ('var, 'value) t -> 'var -> 'value Read.t

val alloc : ('var, 'value) t -> 'var Alloc.t

val check : ('var, 'value) t -> 'var -> unit

Basic instances:

val unit : (unit, unit) t

val field : (Field.t, field) t

Common constructors:

val tuple2 : ('var1, 'value1) t -> ('var2, 'value2) t -> ('var1 * 'var2, 'value1 * 'value2) t

val (*) : ('var1, 'value1) t -> ('var2, 'value2) t -> ('var1 * 'var2, 'value1 * 'value2) t

synonym for tuple2

val tuple3 : ('var1, 'value1) t -> ('var2, 'value2) t -> ('var3, 'value3) t -> ('var1 * 'var2 * 'var3, 'value1 * 'value2 * 'value3) t

val list : length:int -> ('var, 'value) t -> ('var list, 'value list) t

val array : length:int -> ('var, 'value) t -> ('var array, 'value array) t

val hlist : (unit, unit, 'k_var, 'k_value) Data_spec.t -> ((unit, 'k_var) H_list.t, (unit, 'k_value) H_list.t) t

Unpack a Data_spec.t list to a t. The return value relates a polymorphic list of OCaml types to a polymorphic list of R1CS types.

Convert relationships over isomorphic types:

val transport : ('var, 'value1) t -> there:('value2 -> 'value1) -> back:('value1 -> 'value2) -> ('var, 'value2) t

val transport_var : ('var1, 'value) t -> there:('var2 -> 'var1) -> back:('var1 -> 'var2) -> ('var2, 'value) t

val of_hlistable : (unit, unit, 'k_var, 'k_value) Data_spec.t -> var_to_hlist:('var -> (unit, 'k_var) H_list.t) -> var_of_hlist:((unit, 'k_var) H_list.t -> 'var) -> value_to_hlist:('value -> (unit, 'k_value) H_list.t) -> value_of_hlist:((unit, 'k_value) H_list.t -> 'value) -> ('var, 'value) t

module Internal : sig ... end

Typ.ts that make it easier to write a Typ.t for a mix of R1CS data and normal OCaml data.

module Of_traversable (T : Snarky_backendless.Traversable.S) : sig ... end

module type S = Snarky_backendless.Snark_intf.Typ0.Intf.S with type field := Field.Constant.t and type field_var := Field.t and type _ checked = unit

val mk_typ : (module S with type Value.t = 'value and type Var.t = 'var) -> ('var, 'value) t