Module Inputs_intf.Impl

include Snarky_backendless.Snark_intf.Run_basic with type prover_state = unit

type prover_state = unit

The type of state that As_prover blocks may read to/write from.

module Proving_key : sig ... end

The Backend_intf.S.Proving_key module from the backend.

module Verification_key : sig ... end

The Backend_intf.S.Verification_key module from the backend.

module R1CS_constraint_system : sig ... end

The rank-1 constraint system used by this instance. See Backend_intf.S.R1CS_constraint_system.

module Keypair : sig ... end

Managing and generating pairs of keys Proving_key.t and Verification_key.t.

module Var : sig ... end

Variables in the R1CS.

type field

The finite field over which the R1CS operates.

module Bigint : sig ... end

module Constraint : sig ... end

Rank-1 constraints over Field.ts.

module Data_spec : sig ... end

The data specification for checked computations.

module Typ : sig ... end

Mappings from OCaml types to R1CS variables and constraints.

module Boolean : sig ... end

Representation of booleans within a field.

module Field : sig ... end

module As_prover : sig ... end

The functions in this module may only be run as the prover; trying to run them outside of functions that refer to As_prover.t will result in a runtime error.

module Proof_inputs : sig ... end

module Bitstring_checked : sig ... end

module Handle : sig ... end

type response = Snarky_backendless.Request.response

val unhandled : response

type request = Snarky_backendless.Request.request =

| With : {

request : 'a Snarky_backendless.Request.t;

respond : 'a Snarky_backendless.Request.Response.t -> response;

} -> request

module Handler : sig ... end

module Proof_system : sig ... end

val assert_ : ?label:string -> Constraint.t -> unit

val assert_all : ?label:string -> Constraint.t list -> unit

val assert_r1cs : ?label:string -> Field.t -> Field.t -> Field.t -> unit

val assert_square : ?label:string -> Field.t -> Field.t -> unit

val as_prover : (unit -> unit) As_prover.t -> unit

val next_auxiliary : unit -> int

val request_witness : ('var, 'value) Typ.t -> (unit -> 'value Snarky_backendless.Request.t) As_prover.t -> 'var

val perform : (unit -> unit Snarky_backendless.Request.t) As_prover.t -> unit

val request : ?such_that:('var -> unit) -> ('var, 'value) Typ.t -> 'value Snarky_backendless.Request.t -> 'var

TODO: Come up with a better name for this in relation to the above

val exists : ?request:(unit -> 'value Snarky_backendless.Request.t) As_prover.t -> ?compute:(unit -> 'value) As_prover.t -> ('var, 'value) Typ.t -> 'var

val exists_handle : ?request:(unit -> 'value Snarky_backendless.Request.t) As_prover.t -> ?compute:(unit -> 'value) As_prover.t -> ('var, 'value) Typ.t -> ('var, 'value) Handle.t

val handle : (unit -> 'a) -> Handler.t -> 'a

val handle_as_prover : (unit -> 'a) -> (unit -> Handler.t As_prover.t) -> 'a

val if_ : Boolean.var -> typ:('var, _) Typ.t -> then_:'var -> else_:'var -> 'var

if_ b ~then_ ~else_ returns then_ if b is true, or else_ otherwise.

WARNING: The Typ.t's read field must be able to construct values from a series of field zeros.

val with_label : string -> (unit -> 'a) -> 'a

val make_checked : (unit -> 'a) -> ('a, prover_state, field) Snarky_backendless.Types.Checked.t

val constraint_system : exposing:(unit -> 'a, _, 'k_var, _) Data_spec.t -> 'k_var -> R1CS_constraint_system.t

val generate_keypair : exposing:(unit -> 'a, _, 'k_var, _) Data_spec.t -> 'k_var -> Keypair.t

val generate_witness : (unit -> 'a, Proof_inputs.t, 'k_var, 'k_value) Data_spec.t -> 'k_var -> prover_state -> 'k_value

val generate_public_input : (_, Field.Constant.Vector.t, _, 'k_value) Data_spec.t -> 'k_value

Generate the public input vector for a given statement.

val generate_witness_conv : f:(Proof_inputs.t -> 'out) -> (unit -> 'a, 'out, 'k_var, 'k_value) Data_spec.t -> 'k_var -> prover_state -> 'k_value

val run_unchecked : (unit -> 'a) -> prover_state -> prover_state * 'a

val run_and_check : (unit -> (unit -> 'a) As_prover.t) -> prover_state -> (prover_state * 'a) Core_kernel.Or_error.t

val check : (unit -> 'a) -> prover_state -> unit Core_kernel.Or_error.t

val constraint_count : ?weight:(Constraint.t -> int) -> ?log:(?start:bool -> string -> int -> unit) -> (unit -> 'a) -> int

val set_constraint_logger : (?at_label_boundary:([ `Start | `End ] * string) -> Constraint.t -> unit) -> unit

val clear_constraint_logger : unit -> unit

module Internal_Basic : Snarky_backendless.Snark_intf.Basic with type field = field and type 'a As_prover.Ref.t = 'a As_prover.Ref.t

val run_checked : ('a, prover_state) Internal_Basic.Checked.t -> 'a

module Number : Snarky_backendless.Number_intf.Run with type field := field and type field_var := Field.t and type bool_var := Boolean.var

module Enumerable (M : sig ... end) : Snarky_backendless.Enumerable_intf.Run with type ('a, 'b) typ := ('a, 'b) Typ.t and type bool_var := Boolean.var and type var = Field.t and type t := M.t