Module Typ.Alloc

A 'value Alloc.t describes allocating variables for the R1CS to use without explicitly giving the values to store in them. This is analogous to Store.t.

The main use of this is in generating the constraint system and generating the Keypair.t with generate_keypair; we can't know yet what values we will want to store in the variables, but we still want to know what constraints they will have to satisfy.

include Snarky_backendless.Monad_let.S with type 'a t = ('a, Field.t) Snarky_backendless.Typ_monads.Alloc.t

type 'a t = ('a, Field.t) Snarky_backendless.Typ_monads.Alloc.t

include Core_kernel.Monad.S_without_syntax with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val all_ignore : unit t list -> unit t

module Let_syntax : sig ... end

val alloc : Field.Var.t t

Allocate a variable in the R1CS that can hold a single field element.