Module Impl.Bitstring_checked

type t = Boolean.var list

val equal : t -> t -> Boolean.var

val equal_expect_true : t -> t -> Boolean.var

Equivalent to equal, but avoids computing field elements to represent chunks of the list when not necessary.

NOTE: This will do extra (wasted) work before falling back to the behaviour of equal when the values are not equal.

val lt_value : Boolean.var Bitstring_lib.Bitstring.Msb_first.t -> bool Bitstring_lib.Bitstring.Msb_first.t -> Boolean.var

module Assert : sig ... end