Module Internal_Basic.R1CS_constraint_system

The rank-1 constraint system used by this instance. See Backend_intf.S.R1CS_constraint_system.

type t

val digest : t -> Core_kernel.Md5.t

val to_json : t -> 'a Snarky_backendless.Snark_intf.json

Convert a basic constraint into a JSON representation.

This representation is compatible with the Yojson library, which can be used to print JSON to the screen, write it to a file, etc.